api: improve previous commit
If device exists on server and not in JSON, its probably a recently added official device, or there
is the possibility that the maintainer forgot to add the device to the JSON. Don't say that there
aren't any builds availble, because we do have builds. Just display the latest build and the links,
without the maintainer/phone/XDA info.

Signed-off-by: Akhil Narang <akhilnarang.1999@gmail.com>